refactor: route Anthropic messages through InferenceProvider instead of hardcoded provider paths
Extract Anthropic-to-OpenAI translation utilities to shared module under
src/ogx/providers/utils/inference/anthropic_translation.py. Add
anthropic_messages() and anthropic_count_tokens() methods to the
InferenceProvider protocol with a default translation implementation in
OpenAIMixin so every OpenAI-compatible provider gains Anthropic support
automatically. Ollama and vLLM adapters override with native /v1/messages
HTTP passthrough. InferenceRouter dispatches based on the target model. The
messages provider is slimmed to a thin delegator that forwards to the
inference API, retaining only message batch operations that require local
state management.

- Add anthropic_messages() and anthropic_count_tokens() to InferenceProvider protocol
- Add default translation implementation to OpenAIMixin using shared utilities
- Add native passthrough override to OllamaInferenceAdapter and VLLMInferenceAdapter
- Add anthropic_messages() and anthropic_count_tokens() to InferenceRouter
- Slim messages/impl.py to delegate to inference API
- Fix _finalize_batch_error to preserve partial results
- Update error messages to use 'Failed to ...' prefix
- Add missing anthropic methods to SentenceTransformersInferenceImpl
- Update provider registry description to reflect delegation model
- Move translation tests to test_anthropic_translation.py under providers/utils/inference/
- Update test_impl.py to test delegation and batch operations
